chore(client): move logo to assets folder
