+++ client/src/assets/logo/DocTrack-48x48.png
