+++ client/src/assets/logo/DocTrack-72x72.png
