+++ client/src/assets/logo/LogoSidedrawer.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import './../icons/icon-styles.css';
+    import './../../components/icons/icon-styles.css';
     export let alt: string;
 
     const iconUrl = new URL('../../assets/icons/LogoMain.svg', import.meta.url);
