+++ client/src/assets/logo/MainLogo.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
     import { fade } from 'svelte/transition';
-
     
-    import './../icons/icon-styles.css';
+    import './../../components/icons/icon-styles.css';
     export let alt: string;
 
     const iconUrl = new URL('../../assets/icons/LogoMain.svg', import.meta.url);
