+++ client/src/components/ui/navigationbar/SideDrawer.svelte
@@ -17,7 +17,7 @@
     import ChartClusterBar from '../../icons/ChartClusterBar.svelte';
     import OfficeSelect from '../OfficeSelect.svelte';
 
-    import LogoSidedrawer from '../../logo/LogoSidedrawer.svelte';
+    import LogoSidedrawer from '../../../assets/logo/LogoSidedrawer.svelte';
 
     export let show = false;
