+++ client/src/pages/dashboard/index.html
@@ -5,7 +5,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <meta name="referrer" content="no-referrer" />
         <link rel="manifest" href="../manifest.json" />
-        <link rel="icon" href="../../components/logo/DocTrack-256x256.png" />
+        <link rel="icon" href="../../assets/logo/DocTrack-256x256.png" />
         <link rel="stylesheet" href="../global.css" />
         <script src="index.ts" type="module"></script>
         <title>Dashboard</title>
