+++ client/src/pages/manifest.json
@@ -4,22 +4,22 @@
     "description": "A document tracking system built for the modern age.",
     "icons": [
         {
-            "src": "../components/logo/DocTrack-16x16.png",
+            "src": "../assets/logo/DocTrack-16x16.png",
             "sizes": "16x16",
             "type": "image/png"
         },
         {
-            "src": "../components/logo/DocTrack-32x32.png",
+            "src": "../assets/logo/DocTrack-32x32.png",
             "sizes": "32x32",
             "type": "image/png"
         },
         {
-            "src": "../components/logo/DocTrack-72x72.png",
+            "src": "../assets/logo/DocTrack-72x72.png",
             "sizes": "72x72",
             "type": "image/png"
         },
         {
-            "src": "../components/logo/DocTrack-256x256.png",
+            "src": "../assets/logo/DocTrack-256x256.png",
             "sizes": "256x256",
             "type": "image/png"
         }
